create modules and update  functionalities
