+++ package.json
@@ -35,7 +35,7 @@
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.27.5",
     "gh-pages": "^5.0.0",
-    "hint": "^7.1.8",
+    "hint": "^2.0.0",
     "html-webpack-plugin": "^5.5.1",
     "jest-environment-jsdom": "^29.5.0",
     "jest-localstorage-mock": "^2.4.26",
